pyproject.toml
[tool.poetry]
name = "m-wave"
version = "0.1.0"
description = "Mycelium-based Earthquake Detection System"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.0.1"
pyserial = "^3.5"
requests = "^2.26.0"
beautifulsoup4 = "^4.9.3"
numpy = "^1.21.0"
pandas = "^1.3.0"
scikit-learn = "^0.24.2"
python-dotenv = "^0.19.0"
gunicorn = "^20.1.0"
flask-cors = "^3.0.10"
azure-functions = "^1.12.0"
boto3 = "^1.26.0"
google-cloud-storage = "^2.5.0"
openai = "^0.27.0"
anthropic = "^0.3.0"
azure-iot-device = "^2.11.0"
azure-storage-blob = "^12.19.0"
azure-cosmos = "^4.5.1"

[tool.poetry.group.dev.dependencies]
pytest = "^6.2.5"
black = "^21.6b0"
flake8 = "^3.9.2"
mypy = "^0.910"
pytest-cov = "^2.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3 